feat: make last history fields and group fields match

+++ backend/apps/group/api_views.py
@@ -1,3 +1,4 @@
+import requests
 from django.conf import settings
 from django.db.models import Count, F, Q, QuerySet
 from django.http.request import QueryDict
@@ -6,8 +7,6 @@ from django.template.loader import render_to_string
 from django.utils import timezone
 from django.utils.dateparse import parse_datetime
 from django.utils.translation import gettext as _
-
-import requests
 from requests import Request
 from rest_framework import (
     decorators,
@@ -25,7 +24,6 @@ from rest_framework.settings import api_settings
 
 from apps.utils.discord import respond_admin_request, send_admin_request
 from apps.utils.parse import parse_bool
-
 from .models import Group, GroupType, Label, Membership
 from .permissions import (
     AdminRequestListPermission,
@@ -59,7 +57,6 @@ class GeoJsonRender(renderers.JSONRenderer):
 
 class BadRequestException(exceptions.APIException):
     status_code = status.HTTP_400_BAD_REQUEST
-    default_detail = _("Bad request.")
     default_code = "bad_request"
 
 
@@ -327,6 +324,12 @@ class HistoryGroupViewSet(
                 code="invalid",
             )
 
+        if group and group.history.most_recent().history_id == history.history_id:
+            raise BadRequestException(
+                _("You cannot delete the most recent version of a group."),
+                code="invalid",
+            )
+
         history.delete()
         return response.Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -482,7 +485,7 @@ class MembershipViewSet(viewsets.ModelViewSet):
         )
         # check that memberships are from same group
         if (lower and lower.group != member.group) or self.query_params.get(
-            "group"
+                "group"
         ) != member.group.slug:
             raise exceptions.ValidationError(
                 _("All memberships objects must be from the same group."),
@@ -607,8 +610,8 @@ class MembershipViewSet(viewsets.ModelViewSet):
             )
             membership.user.email_user(
                 subject=(
-                    _("Your admin request for %(group)s has been accepted.")
-                    % {"group": membership.group.name}
+                        _("Your admin request for %(group)s has been accepted.")
+                        % {"group": membership.group.name}
                 ),
                 message=mail,
                 html_message=mail,
@@ -628,7 +631,7 @@ class MembershipViewSet(viewsets.ModelViewSet):
         return response.Response(
             {
                 "message": _("The user %(user)s is now admin!")
-                % {"user": membership.user}
+                           % {"user": membership.user}
             },
             status=status.HTTP_202_ACCEPTED,
         )
@@ -658,8 +661,8 @@ class MembershipViewSet(viewsets.ModelViewSet):
             )
             membership.user.email_user(
                 subject=(
-                    _("Your admin request for %(group)s has been denied.")
-                    % {"group": membership.group.name}
+                        _("Your admin request for %(group)s has been denied.")
+                        % {"group": membership.group.name}
                 ),
                 message=mail,
                 html_message=mail,
@@ -679,7 +682,7 @@ class MembershipViewSet(viewsets.ModelViewSet):
         return response.Response(
             {
                 "message": _("The admin request from %(user)s has been denied.")
-                % {"user": membership.user}
+                           % {"user": membership.user}
             },
             status=status.HTTP_202_ACCEPTED,
         )
